rename dummydata to sampleData
