+++ .eslintignore
@@ -1,5 +1,5 @@
 webpack.config.js
 node_modules
 bundle.js
-dummydata.v2.js
+sampleData.js
 coverage/
